+++ NEWS.md
@@ -13,6 +13,8 @@
 
 ## Internal
 
+- re-factoring of vector-raster zonal statistic engines (#150)
+
 - extensive re-writting of testing infrastructure for indicator functions omitting
 the usage of snapshot tests as far as possible (#142)